+++ integration_cdf.py
@@ -1,3 +1,15 @@
-from .tomni.cytosmart_data_format import CytoSmartDataFormat
+#%%
+from tomni.cytosmart_data_format.annotations import Ellipse, Point
 
-dummy = CytoSmartDataFormat([{}])
+# Ellipse annotation
+elli = Ellipse(
+    radius=Point(1, 3),
+    center=Point(0, 0),
+    rotation=0,
+    id="Hello Jan",
+    label="cell",
+    children=[],
+    parents=[],
+)
+
+print(elli.to_dict())