+++ tomni/__init__.py
@@ -7,13 +7,15 @@ __version__ = "1.13.0"
 
 from . import (
     bbox_operations,
-    cytosmart_data_format,
     illumination_correction,
     json_operations,
     make_mask,
     transformers,
 )
 from .bbox_fitting import bbox_fitting, bbox_fitting_center
+from .cytosmart_data_format import CytoSmartDataFormat
+from .cytosmart_data_format.annotations.ellipse import Ellipse
+from .cytosmart_data_format.annotations.point import Point
 from .img_dim import img_dim
 from .img_paste import img_paste
 from .shape_fitting import fit_rect_around_ellipse